Add a ScanCursor wrapper for IterateMainStore() and IterateObjectStore() to avoid tempKV Remove a couple unused Iterate functions Fix reversed Assert.AreEqual params in a test
